refactor(community-client): decompose 837-line monolith into focused _components
