+++ apps/website/app/communities/[slug]/_components/constants.ts
@@ -1,0 +1,15 @@
+export const RANK_COLORS: Record<string, string> = {
+  D: 'border-emerald-500/30 bg-emerald-500/10 text-emerald-400',
+  C: 'border-blue-500/30 bg-blue-500/10 text-blue-400',
+  B: 'border-amber-500/30 bg-amber-500/10 text-amber-400',
+  A: 'border-orange-500/30 bg-orange-500/10 text-orange-400',
+  S: 'border-rose-500/30 bg-rose-500/10 text-rose-400',
+};
+
+export const RANK_TOOLTIPS: Record<string, string> = {
+  D: 'Beginner friendly - Great for getting started',
+  C: 'Intermediate level - Some experience recommended',
+  B: 'Advanced - Requires solid understanding',
+  A: 'Expert level - Significant experience needed',
+  S: 'Only Chonin recommended to contribute - Master level project',
+};
