+++ apps/website/app/communities/[slug]/_components/types.ts
@@ -1,0 +1,1 @@
+export type SortOption = 'rank' | 'latest' | 'stars' | 'name';
