+++ apps/website/app/mentors/mentors-client.tsx
@@ -9,45 +9,7 @@ import Navbar from '../../components/Navbar';
 import SpotlightCard from '../../components/SpotlightCard';
 import { getActiveMentors, type Mentor } from '../../lib/api-client';
 import { COMMUNITIES, SITE_CONFIG } from '../../lib/data';
-
-const REQUIREMENTS = [
-  {
-    icon: '🔥',
-    title: 'Outstanding Open-Source Impact',
-    description:
-      "We value meaningful contributions over years of experience. Show us projects you've impacted, PRs that made a difference, or communities you've helped grow.",
-  },
-  {
-    icon: '⏰',
-    title: 'At Least 4 Hours Per Week',
-    description:
-      'Commit to volunteering at least 4 hours weekly to guide hackers, review code, and participate in community activities.',
-  },
-  {
-    icon: '🎯',
-    title: 'Teaching & Communication',
-    description:
-      'Ability to explain complex concepts clearly, provide constructive feedback, and adapt your mentoring style to different skill levels.',
-  },
-  {
-    icon: '🤝',
-    title: 'Collaborative Spirit',
-    description:
-      'Enthusiasm for working with developers from diverse backgrounds and fostering an inclusive, supportive learning environment.',
-  },
-  {
-    icon: '📋',
-    title: 'Mission Assignment Skills',
-    description:
-      'Capability to identify appropriate challenges for hackers, break down complex projects into achievable missions, and track progress.',
-  },
-  {
-    icon: '🌟',
-    title: 'Any Tech Stack Welcome',
-    description:
-      "We welcome mentors from all technical backgrounds. We'll match you with communities and hackers based on your expertise.",
-  },
-];
+import { REQUIREMENTS } from './_data';
 
 export default function MentorsClient() {
   const applyRef = useRef<HTMLDivElement>(null);
